addBook関数内にバリデーション
